ingestion/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

CHROMA_HOST = os.getenv("CHROMA_HOST", "chromadb")
CHROMA_PORT = int(os.getenv("CHROMA_PORT", 8004))
NEO4J_URI = os.getenv("NEO4J_URI", "bolt://neo4j:7687")
NEO4J_USER = os.getenv("NEO4J_USER", "neo4j")
NEO4J_PASSWORD = os.getenv("NEO4J_PASSWORD", "password")
SHARED_DATA_PATH = os.getenv("SHARED_DATA_PATH", "/shared_data")
EMBEDDING_MODEL = "BAAI/bge-small-en-v1.5"
