ingestion/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
spacy==3.7.4
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.0/en_core_web_lg-3.7.0-py3-none-any.whl
chromadb==0.4.24
sentence-transformers==2.7.0
neo4j==5.19.0
pdfplumber==0.11.0
